restrict pending deposit index to uint32 (#2670)
* Restricted the pending deposit index to uint32 as the pending deposits container height is only 28

* Added check that pendingDepositIndex is < 2**27 in verifyPendingDeposit

+++ contracts/contracts/beacon/BeaconProofs.sol
@@ -115,7 +115,7 @@ contract BeaconProofs is IBeaconProofs {
         bytes32 pendingDepositsContainerRoot,
         bytes32 pendingDepositRoot,
         bytes calldata proof,
-        uint64 pendingDepositIndex
+        uint32 pendingDepositIndex
     ) external view {
         BeaconProofsLib.verifyPendingDeposit(
             pendingDepositsContainerRoot,
